+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `summars-sort-by`: now supports reverse sorting by using a `-` prefix, e.g. `-ALIAS`
+- `summars-json`: new boolean option to output data in json format
 
 ### Fixed